feat(cart): enable non-login user to use session cart

+++ app/Http/Controllers/CartsController.php
@@ -7,20 +7,38 @@ use App\Models\Cart;
 use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\App;
+use Illuminate\Support\Facades\Session;
 
 class CartsController extends Controller
 {
     // GET 购物车清单
     public function index(Request $request)
     {
-        $carts = $request->user()->carts()->with('sku.product')->get();
+        $user = $request->user();
+        if ($user) {
+            $carts = $user->carts()->with('sku.product')->get();
+        } else {
+            $carts = session('carts', []);
+            // $carts = Session::get('carts', []);
+        }
 
         // 自动清除失效商品[已删除或已下架商品]
-        foreach ($carts as $cart) {
+        $flag = false;
+        foreach ($carts as $key => $cart) {
             if (!$cart->sku->product || !$cart->sku->product->on_sale) {
-                $cart->user()->dissociate();
-                $cart->delete();
+                if ($user) {
+                    $cart->user()->dissociate();
+                    $cart->delete();
+                } else {
+                    unset($carts[$key]);
+                    $flag = true;
+                }
             }
+        }
+        if ($flag) {
+            session(['carts' => $carts]);
+            // Session::put('carts', $carts);
+            // Session::put(['carts' => $carts]);
         }
 
         return view('carts.index', [
@@ -31,90 +49,99 @@ class CartsController extends Controller
     // POST 加入购物车
     public function store(CartRequest $request)
     {
-        $cart = Cart::where([
-            'user_id' => $request->user()->id,
-            'product_sku_id' => $request->input('sku_id'),
-        ])->first();
+        $user = $request->user();
+        $sku_id = $request->input('sku_id');
+        $number = $request->input('number');
 
-        if ($cart) {
-            $cart->increment('number', $request->input('number'));
-            $result = $cart->save();
+        if ($user) {
+            $cart = Cart::where([
+                'user_id' => $user->id,
+                'product_sku_id' => $sku_id
+            ])->first();
+            if ($cart) {
+                $cart->increment('number', $number);
+                $cart->save();
+            } else {
+                Cart::create([
+                    'user_id' => $user->id,
+                    'product_sku_id' => $sku_id,
+                    'number' => $number
+                ]);
+            }
         } else {
-            $result = Cart::create([
-                'user_id' => $request->user()->id,
-                'product_sku_id' => $request->input('sku_id'),
-                'number' => $request->input('number'),
-            ]);
+            $carts = session('carts', []);
+            // $carts = Session::get('carts', []);
+            $flag = false;
+            foreach ($carts as $key => $cart) {
+                if ($cart['product_sku_id'] == $sku_id) {
+                    $carts[$key]['number'] += $number;
+                    $flag = true;
+                    break;
+                }
+            }
+            if (!$flag) {
+                $carts[] = [
+                    'product_sku_id' => $sku_id,
+                    'number' => $number
+                ];
+            }
+            session(['carts' => $carts]);
+            // Session::put('carts', $carts);
+            // Session::put(['carts' => $carts]);
         }
 
-        if ($result) {
-            return response()->json([
-                'code' => 200,
-                'message' => 'success',
-            ]);
-        } else {
-            return response()->json([
-                'code' => 422,
-                'message' => 'Unprocessable Entity',
-            ], 422);
-        }
-    }
-
-    // POST 加入购物车
-    public function storeBySkuParameters(CartRequest $request)
-    {
-        $base_size = $request->input('base_size');
-        $hair_colour = $request->input('hair_colour');
-        $hair_density = $request->input('hair_density');
-        $product = Product::find($request->input('product_id'));
-        $skus = $product->skus();
-        if (App::isLocale('zh-CN')) {
-            $skus = $product->is_base_size_optional ? $skus->where('base_size_zh', $base_size) : $skus;
-            $skus = $product->is_hair_colour_optional ? $skus->where('hair_colour_zh', $hair_colour) : $skus;
-            $skus = $product->is_hair_density_optional ? $skus->where('hair_density_zh', $hair_density) : $skus;
-        } else {
-            $skus = $product->is_base_size_optional ? $skus->where('base_size_en', $base_size) : $skus;
-            $skus = $product->is_hair_colour_optional ? $skus->where('hair_colour_en', $hair_colour) : $skus;
-            $skus = $product->is_hair_density_optional ? $skus->where('hair_density_en', $hair_density) : $skus;
-        }
-        $sku = $skus->first();
-        $cart = Cart::where([
-            'user_id' => $request->user()->id,
-            'product_sku_id' => $sku->id,
-        ])->first();
-
-        if ($cart) {
-            $cart->increment('number', $request->input('number'));
-            $result = $cart->save();
-        } else {
-            $result = Cart::create([
-                'user_id' => $request->user()->id,
-                'product_sku_id' => $request->input('sku_id'),
-                'number' => $request->input('number'),
-            ]);
-        }
-
-        if ($result) {
-            return response()->json([
-                'code' => 200,
-                'message' => 'success',
-            ]);
-        } else {
-            return response()->json([
-                'code' => 422,
-                'message' => 'Unprocessable Entity',
-            ], 422);
-        }
+        return response()->json([
+            'code' => 200,
+            'message' => 'success',
+        ]);
     }
 
     // PATCH 更新 (增减数量)
-    public function update(CartRequest $request, Cart $cart)
+    public function update(CartRequest $request)
     {
-        $this->authorize('update', $cart);
+        // $this->authorize('update', $cart);
 
-        $cart->update([
-            'number' => $request->input('number'),
-        ]);
+        $user = $request->user();
+        $sku_id = $request->input('sku_id');
+        $number = $request->input('number');
+
+        if ($user) {
+            $cart = Cart::where([
+                'user_id' => $user->id,
+                'product_sku_id' => $sku_id
+            ])->first();
+            if ($cart) {
+                $cart->update([
+                    'number' => $number
+                ]);
+            } else {
+                Cart::create([
+                    'user_id' => $user->id,
+                    'product_sku_id' => $sku_id,
+                    'number' => $number
+                ]);
+            }
+        } else {
+            $carts = session('carts', []);
+            // $carts = Session::get('carts', []);
+            $flag = false;
+            foreach ($carts as $key => $cart) {
+                if ($cart['product_sku_id'] == $sku_id) {
+                    $carts[$key]['number'] = $number;
+                    $flag = true;
+                    break;
+                }
+            }
+            if (!$flag) {
+                $carts[] = [
+                    'product_sku_id' => $sku_id,
+                    'number' => $number
+                ];
+            }
+            session(['carts' => $carts]);
+            // Session::put('carts', $carts);
+            // Session::put(['carts' => $carts]);
+        }
 
         return response()->json([
             'code' => 200,
@@ -123,12 +150,39 @@ class CartsController extends Controller
     }
 
     // DELETE 删除
-    public function destroy(Request $request, Cart $cart)
+    public function destroy(CartRequest $request)
     {
-        $this->authorize('delete', $cart);
+        // $this->authorize('delete', $cart);
 
-        $cart->user()->dissociate();
-        $cart->delete();
+        $user = $request->user();
+        $sku_id = $request->input('sku_id');
+
+        if ($user) {
+            $cart = Cart::where([
+                'user_id' => $user->id,
+                'product_sku_id' => $sku_id
+            ])->first();
+            if ($cart) {
+                $cart->user()->dissociate();
+                $cart->delete();
+            }
+        } else {
+            $carts = session('carts', []);
+            // $carts = Session::get('carts', []);
+            $flag = false;
+            foreach ($carts as $key => $cart) {
+                if ($cart['product_sku_id'] == $sku_id) {
+                    unset($carts[$key]);
+                    $flag = true;
+                    break;
+                }
+            }
+            if ($flag) {
+                session(['carts' => $carts]);
+                // Session::put('carts', $carts);
+                // Session::put(['carts' => $carts]);
+            }
+        }
 
         return response()->json([
             'code' => 200,
@@ -139,7 +193,16 @@ class CartsController extends Controller
     // DELETE 清空购物车
     public function flush(Request $request)
     {
-        Cart::where(['user_id' => $request->user()->id])->delete();
+        $user = $request->user();
+
+        if ($user) {
+            Cart::where(['user_id' => $user->id])->delete();
+        } else {
+            if (session()->has('carts')) { // (Session::has('carts'))
+                session()->forget('carts');
+                // Session::forget('carts');
+            }
+        }
 
         return response()->json([
             'code' => 200,
